rename: PATCH, DELETE member api를 feature/member로 이동
