+++ src/entities/member/index.ts
@@ -1,10 +1,4 @@
-export {
-  getMember,
-  setMemberClothesThickness,
-  setMemberRegion,
-  setMemberOutingTime,
-  withdrawalAccount,
-} from './api/member';
+export { getMember } from './api/member';
 
 export { storeMember } from './lib/storeMember';
