+++ src/features/member/index.ts
@@ -1,2 +1,10 @@
+export {
+  setMemberGender,
+  setMemberOutingTime,
+  setMemberRegion,
+  setMemberClothesThickness,
+  withdrawalAccount,
+} from './api/member';
+
 export { SelectGender } from './ui/SelectGender/SelectGender';
 export { MemberInformationBar } from './ui/MemberInformationBar/MemberInformationBar';
