+++ src/features/search/ui/GeolocationButton/GeolocationButton.tsx
@@ -2,7 +2,9 @@ import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
-import { setMemberRegion, storeMember } from '@/entities/member';
+import { setMemberRegion } from '@/features/member';
+
+import { storeMember } from '@/entities/member';
 import { regionActions } from '@/entities/region';
 import { SearchPageState } from '@/entities/search';
